Moved helpers and cogs into distinguishing directories
