+++ helpers/EmbedHelper.py
@@ -1,0 +1,57 @@
+# EmbedHelper.py
+
+import os
+from nextcord import Embed, User
+from nextcord.ext import commands
+from dotenv import load_dotenv
+
+load_dotenv()
+
+INV_LINK = os.getenv("INV_LINK")
+
+
+def get_inv_link_embed(bot: commands.Bot, author: User) -> Embed:
+    """
+        Generate the embed containing the Scrier invite link.
+    :param bot:     Bot object this embed is being generated for. Used for retrieving icons: commands.Bot
+    :param author:  User who issued the command triggering this embed: User
+    :return:        nextcord Embed: Embed
+    """
+    embed = Embed(title="DMs Only!", color=0xFFCC33)
+    embed.set_thumbnail(url=bot.user.display_avatar)
+    embed.set_author(name=author.name, icon_url=author.display_avatar)
+
+    return embed
+
+
+def get_dm_info_embed(bot: commands.Bot, author: User) -> Embed:
+    """
+        Generate the embed containing Scrier bot info. Triggered when a user tries to issue Scrier commands in a guild.
+    :param bot:     Bot object this embed is being generated for. Used for retrieving icons: commands.Bot
+    :param author:  User who issued the command triggering this embed: User
+    :return:        nextcord Embed: Embed
+    """
+    embed = Embed(title="Scrier Info", color=0xFF5555)
+    embed.set_thumbnail(url=bot.user.display_avatar)
+    embed.set_author(name=bot.user.name, icon_url=bot.user.display_avatar)
+
+    embed.add_field(name=f"DMs Only", value=f"Scrier allows only DMs to facilitate intro setting and other commands. "
+                                            f"This is to avoid cluttering the text channels of its servers, "
+                                            f"and also to preserve the surprise of a user's shiny new intro.\n",
+                    inline=False)
+    embed.add_field(name="", value="")
+    embed.add_field(name=f"Keeping it Simple", value=f"Scrier's goal is to fulfill its purpose, and only its purpose, "
+                                                     f"as best as possible. Part of meeting this goal is keeping the "
+                                                     f"list of commands supported to the essentials. Use the 'set' "
+                                                     f"command with a link or a search to kick off your intro "
+                                                     f"journey.\n",
+                    inline=False)
+
+    embed.add_field(name=f"Issues?", value=f"Scrier is written by one dude. If you find issues, or have ideas for "
+                                           f"enhancements, please send them to Cream#3465 on Discord, or add an issue "
+                                           f"to the Scrier github: https://github.com/jackmb/Scrier\n",
+                    inline=False)
+    embed.add_field(name="", value="")
+    embed.set_footer(text=f"Generated by {author.name}", icon_url=author.display_avatar.url)
+
+    return embed
